requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
streamlit==1.34.0
sqlalchemy==2.0.30
pytest==8.2.0
requests==2.31.0
python-dotenv==1.0.1
jsonschema==4.22.0
httpx==0.27.0
pandas==2.2.2
google-generativeai==0.5.2
PyYAML==6.0.1
groq==0.4.2
sentence-transformers==2.2.2
